Refactor code and fix not saving new conversations
