+++ utils/app/image-to-text.ts
@@ -1,0 +1,171 @@
+import { Dispatch, MutableRefObject } from 'react';
+
+import { ActionType } from '@/hooks/useCreateReducer';
+
+import { Conversation, Message } from '@/types/chat';
+import { PluginID } from '@/types/plugin';
+import { User } from '@/types/user';
+
+import { HomeInitialState } from '@/pages/api/home/home.state';
+
+import {
+  saveConversation,
+  saveConversations,
+  updateConversationLastUpdatedAtTimeStamp,
+} from './conversation';
+import {
+  removeRedundantTempHtmlString,
+  removeTempHtmlString,
+} from './htmlStringHandler';
+import { removeSecondLastLine } from './ui';
+
+import dayjs from 'dayjs';
+
+interface GetTextByImageProps {
+  imageUrl: string;
+  selectedConversation: Conversation;
+  conversations: Conversation[];
+  homeDispatch: Dispatch<ActionType<HomeInitialState>>;
+  user: User | null;
+  stopConversationRef: MutableRefObject<boolean>;
+}
+
+export async function getTextByImage({
+  imageUrl,
+  selectedConversation,
+  conversations,
+  homeDispatch,
+  user,
+  stopConversationRef,
+}: GetTextByImageProps) {
+  const newMessage: Message = {
+    content: `<img id="${PluginID.IMAGE_TO_TEXT}" src="${imageUrl}" />`,
+    role: 'assistant',
+    pluginId: PluginID.IMAGE_TO_TEXT,
+  };
+  let updatedConversation: Conversation;
+
+  updatedConversation = {
+    ...selectedConversation,
+    messages: [...selectedConversation.messages, newMessage],
+  };
+
+  homeDispatch({ field: 'loading', value: true });
+  homeDispatch({ field: 'messageIsStreaming', value: true });
+
+  const controller = new AbortController();
+
+  // use the imageUrl to call api (image to text)
+  const response = await fetch('/api/image-to-text', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+      'user-token': user?.token || '',
+    },
+    signal: controller.signal,
+    body: JSON.stringify({
+      url: imageUrl,
+    }),
+  });
+  if (!response.ok) {
+    homeDispatch({ field: 'loading', value: false });
+    homeDispatch({ field: 'messageIsStreaming', value: false });
+    throw new Error('Network response was not ok');
+  }
+  const data = response.body;
+  if (!data) {
+    homeDispatch({ field: 'loading', value: false });
+    homeDispatch({ field: 'messageIsStreaming', value: false });
+    return;
+  }
+  // response is ok, continue
+  homeDispatch({ field: 'loading', value: false });
+  const reader = data.getReader();
+  const decoder = new TextDecoder();
+  let done = false;
+  let text = '';
+  let largeContextResponse = false;
+  let showHintForLargeContextResponse = false;
+  const originalMessages =
+    updatedConversation.messages.length > 0
+      ? updatedConversation.messages[updatedConversation.messages.length - 1]
+          .content
+      : '';
+  while (!done) {
+    if (stopConversationRef.current === true) {
+      controller.abort();
+      done = true;
+      break;
+    }
+    const { value, done: doneReading } = await reader.read();
+    done = doneReading;
+    const chunkValue = decoder.decode(value);
+    text += chunkValue;
+
+    if (text.includes('[DONE]')) {
+      text = text.replace('[DONE]', '');
+      done = true;
+    }
+    if (text.includes('[REMOVE_TEMP_HTML]')) {
+      text = removeTempHtmlString(text);
+    }
+
+    if (text.includes('[REMOVE_LAST_LINE]')) {
+      text = text.replace('[REMOVE_LAST_LINE]', '');
+      text = removeSecondLastLine(text);
+    }
+
+    const updatedMessages: Message[] = updatedConversation.messages.map(
+      (message, index) => {
+        if (index === updatedConversation.messages.length - 1) {
+          return {
+            ...message,
+            content:
+              removeTempHtmlString(originalMessages) +
+              removeRedundantTempHtmlString(text),
+            largeContextResponse,
+            showHintForLargeContextResponse,
+          };
+        }
+        return message;
+      },
+    );
+    updatedConversation = {
+      ...updatedConversation,
+      messages: updatedMessages,
+      lastUpdateAtUTC: dayjs().valueOf(),
+    };
+    homeDispatch({
+      field: 'selectedConversation',
+      value: updatedConversation,
+    });
+  }
+  const updatedConversations: Conversation[] = conversations.map(
+    (conversation) => {
+      if (conversation.id === selectedConversation.id) {
+        return updatedConversation;
+      }
+      return conversation;
+    },
+  );
+  // If the conversation is new, add it to the list of conversations
+  if (
+    !updatedConversations.find(
+      (conversation) => conversation.id === updatedConversation.id,
+    )
+  ) {
+    updatedConversations.push(updatedConversation);
+  }
+
+  if (updatedConversations.length === 0) {
+    updatedConversations.push(updatedConversation);
+  }
+  saveConversation(updatedConversation);
+
+  homeDispatch({ field: 'conversations', value: updatedConversations });
+  console.log({ updatedConversations });
+  saveConversations(updatedConversations);
+
+  homeDispatch({ field: 'messageIsStreaming', value: false });
+  updateConversationLastUpdatedAtTimeStamp();
+}
